pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ddf-toolkit"
version = "0.1.0"
description = "Python toolkit for myGEKKO DDF (Device Definition File) artifacts — parse, validate, lint, sign"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "ekon GmbH" },
]
keywords = ["mygekko", "ddf", "iot", "building-automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0",
    "chardet>=5.0",
    "cryptography>=43.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "build>=1.2",
    "types-PyYAML>=6.0",
    "types-chardet>=5.0",
]

[project.scripts]
ddf = "ddf_toolkit.cli:app"

[project.urls]
Homepage = "https://github.com/joeexafionbot-dev/ddf-toolkit"
Repository = "https://github.com/joeexafionbot-dev/ddf-toolkit"
Issues = "https://github.com/joeexafionbot-dev/ddf-toolkit/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ddf_toolkit"]

[tool.ruff]
target-version = "py312"
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]  # line-length handled by formatter

[tool.mypy]
strict = true
python_version = "3.12"
mypy_path = "src"
packages = ["ddf_toolkit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers -ra"

[tool.coverage.run]
source = ["ddf_toolkit"]

[tool.coverage.report]
fail_under = 85
show_missing = true
